+++ Delsoft.Calendars.Belgian/BelgianBaseCalendar.cs
@@ -2,7 +2,7 @@
 
 namespace Delsoft.Calendars.Belgian;
 
-public class BelgianCalendar : BaseCalendar<BelgianHolidaysCalendar>
+public class BelgianBaseCalendar : BaseCalendar<BelgianHolidaysCalendar>
 {
     
 }
