+++ Delsoft.Calendars.Test/Stubs/BaseCalendarStub.cs
@@ -1,0 +1,14 @@
+﻿namespace Delsoft.Calendars.Test.Stubs;
+
+internal class BaseCalendarStub : BaseCalendar<HolidaysCalendarStub>
+{
+    public BaseCalendarStub()
+    {
+    }
+
+    public BaseCalendarStub(int year) 
+        : base(year)
+    {
+        
+    }
+}
