--- Delsoft.Calendars.Test/Stubs/CalendarStub.cs
@@ -1,14 +1,0 @@
-﻿namespace Delsoft.Calendars.Test.Stubs;
-
-internal class CalendarStub : BaseCalendar<HolidaysCalendarStub>
-{
-    public CalendarStub()
-    {
-    }
-
-    public CalendarStub(int year) 
-        : base(year)
-    {
-        
-    }
-}
