+++ Delsoft.Calendars/BaseCalendar.cs
@@ -4,13 +4,7 @@ namespace Delsoft.Calendars;
 
 public abstract class BaseCalendar
 {
-    public static TCalendar Create<TCalendar> (int? year = null)
-        where TCalendar : BaseCalendar, new()
-    {
-        return year == null
-            ? new TCalendar()
-            : (TCalendar) Activator.CreateInstance(typeof(TCalendar), year)!;
-    }
+    
 }
 
 public abstract class BaseCalendar<THolidayCalendar> : BaseCalendar
