+++ Delsoft.Calendars/CalendarFactory.cs
@@ -1,0 +1,12 @@
+namespace Delsoft.Calendars;
+
+public static class CalendarFactory
+{
+    public static TCalendar Create<TCalendar> (int? year = null)
+        where TCalendar : BaseCalendar, new()
+    {
+        return year == null
+            ? new TCalendar()
+            : (TCalendar) Activator.CreateInstance(typeof(TCalendar), year)!;
+    }
+}
